LayoutTests/http/tests/security/mixedContent/data-url-script-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that loads a data URL script.  We should
*not* trigger a mixed content callback because the data URL cannot be corrupted
by an active network attacker.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-data-url-script.html";
></iframe>
</body>
</html>
